refactor: update HttpParser
